refactor(TableGroupHandler): 이벤트 리스너 수정(EventListener -> TransactionalEventListener)
- 패키지 구조 수정
