+++ src/main/java/kitchenpos/tablegroup/application/TableGroupEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.tablegroup.domain;
+package kitchenpos.tablegroup.application;
 
 import java.util.List;
