+++ src/main/java/kitchenpos/tablegroup/application/TableUngroupEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.tablegroup.domain;
+package kitchenpos.tablegroup.application;
 
 public class TableUngroupEvent {
